Open Links To PagerDuty In New Tabs
When embedding a schedule in an iframe, the PagerDuty links are not able
to be opened. This is because PagerDuty sets the 'X-Frame-Options'
header to 'sameorigin', which effectively stops the browser from rending
the page in a iframe.

As such, it's more appropriate to open these links in a new tab.